refactor(image_utils & kmeans): Moved K-Means related functions from image_utils.cpp to kmeans.cpp. Moved include statements to .h files. Created kmeans.h. Improved comments in image_utils.cpp/quantize().
